# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BR0 PR0 BOT Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  user_joined: 'user_joined',
  user_approved: 'user_approved',
  user_declined: 'user_declined',
  media_submitted: 'media_submitted',
  broadcast_sent: 'broadcast_sent',
} as const;

export interface ActivityItem {
  type: ActivityItemType;
  message: string;
  timestamp: string;
  /** @nullable */
  userId?: number | null;
}

export interface DashboardStats {
  totalUsers: number;
  pendingUsers: number;
  approvedUsers: number;
  bannedUsers: number;
  totalMedia: number;
  totalBroadcasts: number;
  recentActivity?: ActivityItem[];
}

export interface PendingUser {
  id: number;
  telegramId: string;
  /** @nullable */
  username: string | null;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  mediaCount: number;
  isAdmin?: boolean;
  submittedAt: string;
}

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  pending: 'pending',
  approved: 'approved',
  declined: 'declined',
  banned: 'banned',
} as const;

export interface User {
  id: number;
  telegramId: string;
  /** @nullable */
  username?: string | null;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  status: UserStatus;
  isAdmin?: boolean;
  mediaCount?: number;
  createdAt: string;
  updatedAt?: string;
}

export type UserDetailStatus = typeof UserDetailStatus[keyof typeof UserDetailStatus];


export const UserDetailStatus = {
  pending: 'pending',
  approved: 'approved',
  declined: 'declined',
  banned: 'banned',
} as const;

export type MediaItemType = typeof MediaItemType[keyof typeof MediaItemType];


export const MediaItemType = {
  photo: 'photo',
  video: 'video',
} as const;

export interface MediaItem {
  id: number;
  userId: number;
  type: MediaItemType;
  telegramFileId: string;
  /** @nullable */
  caption?: string | null;
  createdAt: string;
}

export interface UserDetail {
  id: number;
  telegramId: string;
  /** @nullable */
  username?: string | null;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  status: UserDetailStatus;
  isAdmin?: boolean;
  mediaCount?: number;
  createdAt: string;
  updatedAt?: string;
  media?: MediaItem[];
}

export type UserUpdateStatus = typeof UserUpdateStatus[keyof typeof UserUpdateStatus];


export const UserUpdateStatus = {
  approved: 'approved',
  banned: 'banned',
} as const;

export interface UserUpdate {
  status: UserUpdateStatus;
}

export interface Broadcast {
  id: number;
  mediaId: number;
  sentCount: number;
  failedCount: number;
  createdAt: string;
}

export interface Settings {
  id: number;
  botName: string;
  /** @nullable */
  adminPassword?: string | null;
  minMediaRequired: number;
  /** @nullable */
  welcomeMessage?: string | null;
  /** @nullable */
  approvalMessage?: string | null;
  /** @nullable */
  declineMessage?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface SettingsInput {
  /** @minLength 1 */
  botName?: string;
  adminPassword?: string;
  /** @minimum 1 */
  minMediaRequired?: number;
  welcomeMessage?: string;
  approvalMessage?: string;
  declineMessage?: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
/**
 * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
 */
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};