# lib/api-zod/src/generated/types/authUserEnvelope.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BR0 PR0 BOT Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from './authUser';

export interface AuthUserEnvelope {
  user: AuthUser | null;
}